componentize discussed topics
